Implement new CLI
